+++ lib/__init__.py
@@ -3,13 +3,22 @@ import sublime
 
 st_version = int(sublime.version())
 if st_version > 3000:
+    from JoomlaPack.lib.file import File
+    from JoomlaPack.lib.folder import Folder
     from JoomlaPack.lib.helper import Helper
+    from JoomlaPack.lib.json_file import JsonFile
     from JoomlaPack.lib.project import Project
 else:
+    from lib.file import File
+    from lib.folder import Folder
     from lib.helper import Helper
+    from lib.json_file import JsonFile
     from lib.project import Project
 
 __all__ = [
+    'File',
+    'Folder',
     'Helper',
+    'JsonFile',
     'Project'
 ]